Refactor controllers for activity location handling
Moved activity location logic to a new `ActivityLocationController`
to improve separation of concerns. Added endpoints for updating
activity location status and retrieving all activity locations.

Removed activity location-related endpoints and dependencies
from `LocationController`. Updated imports and namespaces
accordingly.
