+++ BackEnd/Peach_ActiviGo.Api/Controllers/LocationController.cs
@@ -1,8 +1,6 @@
 using FluentValidation;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Peach_ActiviGo.Services.DTOs.ActivityLocationDto;
 using Peach_ActiviGo.Services.DTOs.LocationDto;
 using Peach_ActiviGo.Services.Interface;
 
@@ -15,12 +13,10 @@ namespace Peach_ActiviGo.Api.Controllers
     public class LocationController : ControllerBase
     {
         private readonly ILocationService _locationService;
-        private readonly IActivityLocationService _activityLocationService;
 
-        public LocationController(ILocationService locationService, IActivityLocationService activityLocationService)
+        public LocationController(ILocationService locationService)
         {
             _locationService = locationService;
-            _activityLocationService = activityLocationService;
         }
 
         // GET: api/Location
@@ -88,32 +84,6 @@ namespace Peach_ActiviGo.Api.Controllers
                 return NotFound(new { errorMessage = "Location not found!" });
             }
             return Ok("Location updated successfully");
-        }
-
-        [HttpPut("UpdateActivityLocationStatus")]
-        public async Task<IActionResult> UpdateActivityStatusLocation([FromBody] UpdateActivityLocationDto dto, CancellationToken ct, IValidator<UpdateActivityLocationDto> validator)
-        {
-            var validationResult = await validator.ValidateAsync(dto);
-            if (!validationResult.IsValid)
-            {
-                return BadRequest(validationResult.Errors);
-            }
-
-            var result = await _activityLocationService.UpdateActivityLocationStatusAsync(dto, ct);
-            if (!result)
-            {
-                return NotFound(new { errorMessage = $"No ActivityLocation found with Id {dto.id}" });
-            }
-
-            return Ok(new { message = $"ActivityLocation {(dto.IsActive ? "activated" : "deactivated")} successfully." });
-        }
-
-        [HttpGet("GetAllActivityLocations")]
-        public async Task<IActionResult> GetAllActivityLocations(CancellationToken ct)
-        {
-            var locations = await _activityLocationService.GetAllActivityLocationsAsync(ct);
-
-            return Ok(locations);
         }
     }
 }
